added a temp password field and did some refactoring
