+++ README.md
@@ -1,5 +1,14 @@
 # FinanceApp
 Finance app created for practicing full stack app development.
 
-![Finance App Design - Banks@2x](https://user-images.githubusercontent.com/31596996/178000682-49aacfae-880b-402f-9d58-bd9ed8f17f22.png)
+[Main App Directories](https://github.com/Jackal1420/FinanceApp/tree/Add-Dependencies-And-Database/app/src/main/java/com/example/finance)
 
+Directories: 
+    - Data
+        - Data objects that are communicated between the server and client 
+    - DependencyInjection
+        - "Singleton" objects that are provided to the entirety of the app regardless of context  
+    - Ui
+        - Views and ViewModels for individual app screens 
+    - Util
+        - Classes that provide common information to our app (e.g. Standard currency conversion numbers)
