+++ app/src/main/AndroidManifest.xml
@@ -3,7 +3,7 @@
     package="com.example.finance">
 
     <application
-        android:name=".FinanceApp"
+        android:name=".ui.views.FinanceApp"
         android:allowBackup="true"
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
